Unfinished work, added content/config file hierarchy
