+++ GSOCLMapPrototype/GSOCLMapPrototype/GSOCLMapPrototype.vcxproj
@@ -155,19 +155,23 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClInclude Include="ConfigurationSingleton.h" />
     <ClInclude Include="Connection.h" />
     <ClInclude Include="json.h" />
     <ClInclude Include="json.hpp" />
     <ClInclude Include="MapManagerSingleton.h" />
     <ClInclude Include="Map_Node.h" />
+    <ClInclude Include="Map_Node_Old.h" />
     <ClInclude Include="pch.h" />
+    <ClInclude Include="Terrain.h" />
     <ClInclude Include="Tests.h" />
   </ItemGroup>
   <ItemGroup>
+    <ClCompile Include="ConfigurationSingleton.cpp" />
     <ClCompile Include="Export.cpp" />
     <ClCompile Include="Main.cpp" />
     <ClCompile Include="MapManagerSingleton.cpp" />
-    <ClCompile Include="Map_Node.cpp" />
+    <ClCompile Include="Map_Node_Old.cpp" />
     <ClCompile Include="pch.cpp">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">Create</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">Create</PrecompiledHeader>
