+++ GSOCLMapPrototype/GSOCLMapPrototype/Terrain.h
@@ -1,0 +1,6 @@
+#pragma once
+#include <string>
+struct Terrain {
+	std::string name;
+	float* modifiers;
+};
