+++ GSOCLMapPrototype/GSOCLMapPrototype/content/config/config.json
@@ -1,0 +1,4 @@
+{
+  "MAX_NUMBER_OF_NODES": 500000,
+  "MAX_NUM_CONNECTIONS": 20
+}
